# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: SOL Duel Game API
  description: |
    **SOL Duel** is a real-time PvP game built on the Solana blockchain. This API provides endpoints for player authentication, match management, game state synchronization, and blockchain interactions.

    ## Features
    - 🔐 Wallet-based authentication with signature verification
    - ⚔️ Real-time multiplayer combat system
    - 💰 Solana-based wagering and rewards
    - 🏆 Tournament and leaderboard systems
    - 🔗 WebSocket real-time updates
    - 📊 Comprehensive player statistics

    ## Authentication
    This API uses JWT tokens for authentication. Players authenticate using their Solana wallet signature.

    ## Rate Limiting
    - 100 requests per 15 minutes per IP address
    - WebSocket connections limited to 10 per authenticated user

    ## WebSocket Events
    Real-time game events are handled through WebSocket connections. See the WebSocket documentation for details.
  version: 1.0.0
  contact:
    name: SOL Duel API Support
    email: support@solduel.game
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.solduel.game
    description: Production server
  - url: https://staging-api.solduel.game
    description: Staging server
  - url: http://localhost:5000
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /api/games:
    get:
      tags:
        - Games
      summary: Get available games
      description: Retrieve list of available games with filtering and pagination
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [waiting, active, finished]
            default: waiting
          description: Filter games by status
        - name: gameType
          in: query
          schema:
            type: string
            enum: [duel, tournament, practice]
          description: Filter games by type
        - name: minWager
          in: query
          schema:
            type: number
            minimum: 0
          description: Minimum wager amount in SOL
        - name: maxWager
          in: query
          schema:
            type: number
            minimum: 0
          description: Maximum wager amount in SOL
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of games per page
      responses:
        '200':
          description: Games retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      games:
                        type: array
                        items:
                          $ref: '#/components/schemas/Game'
                      pagination:
                        $ref: '#/components/schemas/PaginationInfo'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Games
      summary: Create new game
      description: Create a new game with specified configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
            examples:
              duelGame:
                summary: Standard duel game
                value:
                  gameType: "duel"
                  wagerAmount: 0.1
                  isPrivate: false
                  maxPlayers: 2
                  timeLimit: 300
                  settings: {}
              tournamentGame:
                summary: Tournament game
                value:
                  gameType: "tournament"
                  wagerAmount: 0.5
                  isPrivate: false
                  maxPlayers: 8
                  timeLimit: 600
                  settings:
                    eliminationStyle: "single"
                    bracketsEnabled: true
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Game created successfully"
                  data:
                    type: object
                    properties:
                      game:
                        $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/games/{id}:
    get:
      tags:
        - Games
      summary: Get specific game
      description: Retrieve detailed information about a specific game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Game ID
          example: "game_12345"
      responses:
        '200':
          description: Game details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      game:
                        $ref: '#/components/schemas/GameDetails'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or wallet verification
  schemas:
...
    CreateGameRequest:
      type: object
      required:
        - gameType
        - wagerAmount
      properties:
        gameType:
          type: string
          enum: [duel, tournament, practice]
          description: Type of game
          example: "duel"
        wagerAmount:
          type: number
          minimum: 0
          description: Wager amount in SOL
          example: 0.1
        isPrivate:
          type: boolean
          default: false
          description: Whether game is private
        maxPlayers:
          type: integer
          minimum: 2
          maximum: 8
          default: 2
          description: Maximum number of players
        timeLimit:
          type: integer
          minimum: 60
          maximum: 3600
          default: 300
          description: Time limit in seconds
        settings:
          type: object
          description: Additional game settings
          default: {}
    Game:
      type: object
      properties:
        id:
          type: string
          description: Unique game identifier
          example: "game_12345"
        gameType:
          type: string
          enum: [duel, tournament, practice]
          example: "duel"
        status:
          type: string
          enum: [waiting, active, finished, cancelled]
          example: "waiting"
        wagerAmount:
          type: number
          description: Wager amount in SOL
          example: 0.1
        maxPlayers:
          type: integer
          example: 2
        currentPlayers:
          type: integer
          description: Number of players currently in game
          example: 1
        timeLimit:
          type: integer
          description: Time limit in seconds
          example: 300
        isPrivate:
          type: boolean
          example: false
        creator:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            eloRating:
              type: integer
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
    GameDetails:
      allOf:
        - $ref: '#/components/schemas/Game'
        - type: object
          properties:
            players:
              type: array
              items:
                $ref: '#/components/schemas/GamePlayer'
            currentTurn:
              type: string
              description: Current player's turn (player ID)
            turnTimeRemaining:
              type: integer
              description: Seconds remaining in current turn
            gameState:
              type: object
              description: Current game state data
            rewardPool:
              type: number
              description: Total reward pool in SOL
    GamePlayer:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        eloRating:
          type: integer
        health:
          type: integer
        mana:
          type: integer
        isAlive:
          type: boolean
        isReady:
          type: boolean
        joinedAt:
          type: string
          format: date-time
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 20
        total:
          type: integer
          description: Total number of items
          example: 156
        pages:
          type: integer
          description: Total number of pages
          example: 8

    # Error Schemas
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional error details (optional)
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Validation error in request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Validation Error"
            message: "Invalid request data"
            details: "username must be between 3 and 30 characters"
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication Failed"
            message: "Valid authentication token required"
    ForbiddenError:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Access Denied"
            message: "You do not have permission to perform this action"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not Found"
            message: "The requested resource was not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred"
...